requirements.txt
Django==3.2.8
django-redis==5.0.0
djangorestframework==3.12.4
httplib2==0.14.0
JSON-log-formatter==0.4.0
Pillow==7.0.0
PyJWT==1.7.1
python-dateutil==2.7.3
PyYAML==5.3.1
redis==3.5.3
requests==2.22.0
simplejson==3.16.0
six==1.14.0
sqlparse==0.4.2
urllib3==1.25.8
psycopg2-binary==2.8.5
drf-yasg
easy-thumbnails
django-crispy-forms

